backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
aiosmtplib>=3.0.0
jinja2>=3.1.0
slowapi>=0.1.9
python-dotenv>=1.0.0
python-multipart>=0.0.9

# Dev / Test
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
